Refactor and order of ui activties
