+++ .idea/.name
@@ -1,1 +1,1 @@
-JetpackComposeNewsApp
+VinilosApp
